requirements.txt
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
autobahn==19.11.2
Automat==25.4.16
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.3
Flask==3.1.1
frozenlist==1.6.0
greenlet==3.2.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiteconnect==5.0.1
MarkupSafe==3.0.2
multidict==6.4.4
playwright==1.52.0
propcache==0.3.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyee==13.0.0
pyOpenSSL==25.1.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
questdb==2.0.4
requests==2.32.3
service-identity==24.2.0
six==1.17.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==12.0
Werkzeug==3.1.3
yarl==1.20.0
zope.interface==7.2
